feat: extract decryptStream to its file
